Unit tests passing for Pi4 (finally!)
The main change is integrating BasicSettings into BaseKernel.
This had a shitload of fallout.

The reason fori adjustment this is that it is no more possible to compile
for vc4 and vc6 in parallel. This is because we now have to
take vc7 into account.

vc7 is not enabled yet but the code is making room for it.
An issue is that vc6 and vc7 both fall under v3d. Hoping for the best.