Fixed combine smul24 mul reg a indirection
Yes, I know the title looks like gobbledygook.
Retry:
smul24 runs on the mul ALU and can therefore be combined with a previous add instruction.
When the combine happen, the src a address was lost when using mesa2. This is what I fixed.

OK, this is still gobbledygook. I don't care, I fixed it.
I'm so happy right now, I might even go to church.